update lolipop to use config file instead of crazy args
with this update, a single config module can be used to provide details
on how to connect to the mysql database.  the config is called with
'require' in each nodejs file but is only loaded by node one time.

this provides functionality for connection on each call to a function
within lolipop. rather than passing a connection object around, we can
now pass a config around.

functionality for specifying a database to read in the showTables method
has been removed.  this is now handled by the config.js module and we
will simply assume that the connection is made on a specific database
and that we are only working with that single database.

a test file is added (not an official test set, but rather a test to
make sure that the modules aren't broken)